+++ storybook/src/Overlays/Popovers/Popover.stories.tsx
@@ -74,6 +74,23 @@ const options = [
   { label: 'Pineapples5', value: '45' },
   { label: 'Kiwis5', value: '55' },
 ]
+export const All = () => (
+  <SpaceVertical align="start">
+    <OverlayOpenDialog />
+    <RenderProps />
+    <RenderPropsSpread />
+    <Placement />
+    <PopoverFocusTrap />
+    <Grouped />
+    <MovingTarget />
+    <MouseUp />
+  </SpaceVertical>
+)
+
+export default {
+  component: All,
+  title: 'Overlays/Popover',
+}
 
 export const PopoverFocusTrap = () => {
   function getButtonAlert(text: string) {
@@ -354,6 +371,32 @@ export const MovingTarget = () => {
   )
 }
 
-export default {
-  title: 'Overlays/Popover',
+export const MouseUp = () => {
+  return (
+    <SpaceVertical>
+      <Paragraph>
+        Test that the the 1st click outside is cancelled and that clicking the
+        Popover's trigger a 2nd time closes the Popover
+      </Paragraph>
+      <Paragraph>
+        Previously, the useCapture = true on the mouseup listener caused the
+        click outside behavior to break on any page that has a React onMouseUp
+        on any element.
+      </Paragraph>
+      <Space>
+        <Popover content="Popover 1">
+          <Button>Open 1st Popover</Button>
+        </Popover>
+        <Popover content="Popover 2">
+          <Button>Open 2nd Popover</Button>
+        </Popover>
+        <Button
+          onClick={() => alert('I should be canceled if a Popover was open')}
+        >
+          Click
+        </Button>
+        <Button onMouseUp={() => alert('A simple onMouseUp')}>onMouseUp</Button>
+      </Space>
+    </SpaceVertical>
+  )
 }